Move JSON functions to own file
